remove jquery step 1
