+++ src/website/compare.html
@@ -1,124 +1,25 @@
 <%include file="header.html"/>
 <%include file="nav_menu.html"/>
 
-<div class="tabs is-centered" id="nav">
-  <ul>
-    <li data-target="pane-1" id="1" class="is-active">
-      <a><span>Spinorama</span></a>
-    </li>
-    <li data-target="pane-2" id="2">
-      <a><span>Early Reflections</span></a>
-    </li>
-    <li data-target="pane-3" id="3">
-      <a><span>Estimated In-Room Response</span></a>
-    </li>
-    <li data-target="pane-4" id="4">
-      <a><span>Horizontal Reflections</span></a>
-    </li>
-    <li data-target="pane-5" id="5">
-      <a><span>Vertical Reflections</span></a>
-    </li>
-<!--
-     <li data-target="pane-6" id="6">
-       <a><span>SPL Horizontal</span></a>
-     </li>
-     <li data-target="pane-7" id="7">
-       <a><span>SPL Vertical</span></a>
-     </li>
--->
-  </ul>
-</div>
+<%namespace name="util" module="utils" import="*"/>
 
-<article class="message is-info banner">
-  <div class="message-header">
-    <p>Information</p>
-    <button class="delete" aria-label="delete"></button>
+<div id="cea2034plots" data-num="0">
+  <div class="plotSingle" id="plotSingle"></div>
+  <div class="plotDouble0" id="plotDouble0"></div>
+  <div class="plotDouble1" id="plotDouble1"></div>
+  <div class="plotForm">
+    Graph: <select class="graph"></select></br>
+    <b>A</b>
+    Speaker: <select class="speaker0"></select>
+    Origin: <select class="origin0"></select>
+    Version: <select class="version0"></select></br>
+    <b>B</b>
+    Speaker: <select class="speaker1"></select>
+    Origin <select class="origin1"></select>
+    Version: <select class="version1"></select>
   </div>
-  <div class="message-body">
-    Loading the data takes some time. Please be patient.
-  </div>
-</article>
-
-<div class="tab-pane is-active" id="pane-1">
-  <div id="visCEA2034"></div>
-  <script type="text/javascript">
-    const specCEA2034 = downloadZip("compare/CEA2034.json.zip").then(
-      function(spec2034) {
-        displayGraph(JSON.parse(spec2034), visCEA2034);
-    });
-  </script>
-  <%include file="graph_help.html"/>
 </div>
 
-<div class="tab-pane" id="pane-2">
-  <div id="visER"></div>
-  <script type="text/javascript">
-    const specER = downloadZip("compare/Early Reflections.json.zip").then(
-      function(specER) {
-        displayGraph(JSON.parse(specER), visER);
-    });
-  </script>
-  <%include file="graph_help.html"/>
-</div>
+<script src='./assets/compare.js'></script>
 
-<div class="tab-pane" id="pane-3">
-  <div id="visEIR"></div>
-  <script type="text/javascript">
-    const specEIR = downloadZip("compare/Estimated In-Room Response.json.zip").then(
-      function(specEIR) {
-        displayGraph(JSON.parse(specEIR), visEIR);
-    });
-  </script>
-  <%include file="graph_help.html"/>
-</div>
-
-<div class="tab-pane" id="pane-4">
-  <div id="visHR"></div>
-  <script type="text/javascript">
-    const specHR = downloadZip("compare/Horizontal Reflections.json.zip").then(
-      function(specHR) {
-        displayGraph(JSON.parse(specHR), visHR);
-    });
-  </script>
-  <%include file="graph_help.html"/>
-</div>
-
-<div class="tab-pane" id="pane-5">
-  <div id="visVR"></div>
-  <script type="text/javascript">
-    const specVR = downloadZip("compare/Vertical Reflections.json.zip").then(
-      function(specVR) {
-        displayGraph(JSON.parse(specVR), visVR);
-    });
-  </script>
-  <%include file="graph_help.html"/>
-</div>
-
-<!--
- <div class="tab-pane" id="pane-6">
-   <div id="visSH"></div>
-   <script type="text/javascript">
-     const specSH = downloadZip("compare/SPL Horizontal.json.zip").then(
-       function(specSH) {
-         displayGraph(JSON.parse(specSH), visSH);
-     });
-   </script>
-   <%include file="graph_help.html"/>
- </div>
-
- <div class="tab-pane" id="pane-7">
-   <div id="visSV"></div>
-   <script type="text/javascript">
-     const specSV = downloadZip("compare/SPL Vertical.json.zip").then(
-       function(specSV) {
-         displayGraph(JSON.parse(specSV), visSV);
-     });
-   </script>
-   <%include file="graph_help.html"/>
- </div>
--->
-
-<script src="${site}/assets/tabs.js"></script>
-
-</body>
-</html>
+<%include file="footer.html"/>
