--- src/website/compare2.html
@@ -1,26 +1,0 @@
-<%include file="header.html"/>
-<%include file="nav_menu.html"/>
-
-<%include file="tabs_header.html"/>
-
-<div id="cea2034plots" data-num="0">
-  <div class="plotSingle" id="plotSingle"></div>
-  <div class="plotDouble0" id="plotDouble0"></div>
-  <div class="plotDouble1" id="plotDouble1"></div>
-  <div class="plotForm">
-    Graph: <select class="graph"></select></br>
-    <b>A</b>
-    Speaker: <select class="speaker0"></select>
-    Origin: <select class="origin0"></select>
-    Version: <select class="version0"></select></br>
-    <b>B</b>
-    Speaker: <select class="speaker1"></select>
-    Origin <select class="origin1"></select>
-    Version: <select class="version1"></select>
-  </div>
-</div>
-
-<script src='./assets/compare2.js'></script>
-
-</body>
-</html>
