+++ src/website/footer.html
@@ -1,3 +1,5 @@
+<script src="${site}/assets/bulma.js"></script>
+
 <footer class="footer">
   <div class="content has-text-centered">
     <p>
