+++ src/website/head.html
@@ -2,12 +2,9 @@
   <meta name="viewport" content="width=device-width, initial-scale=1"/>
   <title>A collection of loudspeakers measurements</title>
 
-  <script src="https://cdn.jsdelivr.net/npm/jquery@3.5.0"></script>
   <script src="https://cdn.jsdelivr.net/npm/fuse.js@5.2.1/dist/fuse.min.js"></script>
   <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bulma@0.9.2/css/bulma.min.css" integrity="sha256-O8SsQwDg1R10WnKJNyYgd9J3rlom+YSVcGbEF5RmfFk=" crossorigin="anonymous">
   <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/@fortawesome/fontawesome-free@5.15.2/css/all.min.css" integrity="sha256-0fuNgzfNIlaClbDtmYyFxY8LTNCDrwsNshywr4AALy0=" crossorigin="anonymous">
 
-  <script src="${site}/assets/search.js"></script>
-  <script src="${site}/assets/bulma.js"></script>
   <link crossorigin="anonymous" media="all" rel="stylesheet" href="${site}/assets/spinorama.css" />
   <link rel="icon" type="image/png" href="${site}/assets/favicon-16x16.png" sizes="16x16" />
