+++ src/website/head_vega.html
@@ -1,3 +1,4 @@
+  <script src="${site}/assets/graph.js"></script>
   <script src="${site}/assets/zip.min.js"></script>
   <script src="${site}/assets/downloadzip.js"></script>
   <script src="https://cdn.jsdelivr.net/npm/plotly.js@2.8.0/dist/plotly.min.js" integrity="sha256-Tw6voo+dKwux2fkaNIovxAS2vvXcEz5lTjpA8CjfcBE=" crossorigin="anonymous"></script>
