+++ src/website/nav_menu.html
@@ -11,7 +11,7 @@
     <div class="navbar-end">
       <a class="navbar-item" href="${site}/index.html">Search</a>
       <a class="navbar-item" href="${site}/eqs.html">EQs</a>
-      <a class="navbar-item" href="${site}/compare2.html">Compare</a>
+      <a class="navbar-item" href="${site}/compare.html">Compare</a>
 <!--
       <a class="navbar-item" href="${site}/scores.html">Scores</a>
       <a class="navbar-item" href="${site}/statistics.html">Statistics</a>
