+++ src/website/nav_search.html
@@ -1,27 +1,29 @@
-  <div class="level has-background-info-dark">
-    <div class="level-item">
-      <div class="control has-icons-left">
-        <input id="searchInput" class="input" type="text"
-               placeholder="brand or model ..." autofocus />
-        <div class="icon is-left"><i class="fas fa-search"></i></div>
+<script src="${site}/assets/search.js"></script>
+
+<div class="level has-background-info-dark">
+  <div class="level-item">
+    <div class="control has-icons-left">
+      <input id="searchInput" class="input" type="text"
+             placeholder="brand or model ..." autofocus />
+      <div class="icon is-left"><i class="fas fa-search"></i></div>
+    </div>
+    <div class="dropdown is-hoverable">
+      <div class="dropdown-trigger">
+        <button class="button" aria-haspopup="true" aria-controls="filters-dropdown-menu">
+          <span class="icon">
+            <i class="fas fa-cog" aria-hidden="true"></i>
+          </span>
+        </button>
       </div>
-      <div class="dropdown is-hoverable">
-        <div class="dropdown-trigger">
-          <button class="button" aria-haspopup="true" aria-controls="filters-dropdown-menu">
-            <span class="icon">
-              <i class="fas fa-cog" aria-hidden="true"></i>
-            </span>
-          </button>
-        </div>
-        <div class="dropdown-menu" id="filters-dropdown-menu" role="menu">
-          <div class="dropdown-content">
-            ${filters()}
-            ${sortBy()}
-          </div>
+      <div class="dropdown-menu" id="filters-dropdown-menu" role="menu">
+        <div class="dropdown-content">
+          ${filters()}
+          ${sortBy()}
         </div>
       </div>
     </div>
   </div>
+</div>
 
 <%def name="filters()">
   ${filterQuality()}
